requirements.txt
# --- Core Framework ---
langchain
langchain-community
langchain-core

# --- Vector Database ---
langchain-chroma
chromadb

# --- Models ---
langchain-ollama
langchain-huggingface
sentence-transformers

# --- Document Processing ---
pymupdf
rapidocr-onnxruntime
python-docx
python-pptx
pillow
pandas
openpyxl

# --- Web Server, Security & Monitoring ---
fastapi
uvicorn
python-multipart
pydantic
aiofiles       # Async File I/O
psutil         # Memory Monitoring
slowapi        # Rate Limiting
limits         # Dependency for slowapi